backend/requirements.txt

Flask==2.2.5
discord.py==2.3.2
wavelink==2.6.5
python-dotenv==1.0.1
